Added B-LDPC Codes implementation
B-LDPC Codes implementation in IT++ consists of a BLDPC_Parity and
BLDPC_Generator classes. The former one can be used for constructing
the parity-check matrix from so-called base-matrix.  Base matrix can
be saved and loaded from a text file. The BLDPC_Generator class
implements an efficient encoding algorithm using a preprocessed
version of the parity-check matrix.


git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1062 f17a810d-92cc-4fca-aeb8-6e32867107aa