Use copy_vector() where possible
This patch improves the consistency and possibly the performance
of various Vec's methods by using copy_vector() instead of
explicit copying of elements in a loop. It also fixes a minor bug
in "/=" operator for the case of division of a vector with itself,
i.e. "v /= v;". Finally, this patch adds missing friend
declarations for concat() functions and improves if statements in
the alloc() function.

git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1091 f17a810d-92cc-4fca-aeb8-6e32867107aa