Sorting functions revised
The sort() and sort_index() functions can now use four different
algorithms: Introsort, Quicksort, Heapsort and Insertion Sort, which
are implemented in a common Sort class. The Introsort is the default
method. The revised implementation is much faster because it uses
pointers for accessing the data to be sorted. Sorting test program
is included in this commit as well.

Thanks to Mark Dobossy for this improved implementation.
This commit closes feature request [1746876].


git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1096 f17a810d-92cc-4fca-aeb8-6e32867107aa