Fixed regressions related to the specialised operator*()
The templated friend operator*() for two vectors, which calculates
the dot product, needs to be instantiated in the same way as the
dot() function. Besides, the declaration of the friend operator*()
included the definition. This definition is now moved below dot()
function definition. Without these two changes, a segmentation
fault occurred in vec_test under Cygwin.

git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1127 f17a810d-92cc-4fca-aeb8-6e32867107aa