Improved log2() and log1p() portability
Additional checks for log2() and log1p() functions have been added
to the configure script. If they are not found, local
implementations are used. Besides, log2() is now in a global
scope, not in the "itpp" namespace, which is consistent with GCC.
BTW, removed unnecessary checks for C++ standard header files.

git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1170 f17a810d-92cc-4fca-aeb8-6e32867107aa