Fixed failing blas_test and vec_test when linking with MKL
There was a problem with using zdotusub_ Fortran wrapper with
Intel's MKL library, which resulted in segmentation faults in
blas_test and vec_test programs. Now zdotusub_ is defined as
zdotu_ when MKL BLAS is used, and zdotusub.f is not built in such
a case. BTW, the detection of MKL, ACML and ATLAS BLAS libraries
is improved (additional HAVE_BLAS_MKL, HAVE_BLAS_ACML and
HAVE_BLAS_ATLAS definitions are available).

git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1201 f17a810d-92cc-4fca-aeb8-6e32867107aa