operator*(Vec &v, Mat &m) documented as deprecated
The existence of this operator in IT++ is questionable, since its
behaviour is equivalent to "outer_product(v, m.get_row(0))".
Moreover, the documentation was wrong so many users had problems
using it correctly. This patch marks this operator as deprecated,
both in the documentation and through a run-time warning message.

git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/trunk@1259 a6533b66-300f-0410-b061-b36de6c4fb2d