operator*(Vec &v, Mat &m) documented as deprecated
The existence of this operator in IT++ is questionable, since its
behaviour is equivalent to "outer_product(v, m.get_row(0))".
Moreover, the documentation was wrong so many users had problems
using it correctly. This patch marks this operator as deprecated,
both in the documentation and through a run-time warning message.

git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1259 f17a810d-92cc-4fca-aeb8-6e32867107aa