Bump autoconf requirement to 2.60 and use AC_PROG_SED
The $datarootdir and $docdir variables, which are used for HTML
documentation installation, were introduced in autoconf 2.60.
Therefore we need to check for this version. Besides, manual check
for sed program is now replaced with AC_PROG_SED macro, which
makes sed as a required dependency. Additionally, gdiff and
gnudiff are searched as diff replacement.

git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1273 f17a810d-92cc-4fca-aeb8-6e32867107aa