Fixed wrong calculation of bessely() detected on Intel C++ compiler
When IT++ was compiled with Intel C++ (icpc) compiler, the
bessel_test failed on calculation of bessely() function for
non-integer orders. This bug was caused by the fact that the
ported bessel functions from Cephes Math Library should use
a gamma function, which sets "signgam" global variable. But
instead C99 tgamma() was used, which does not set "signgam".

This patch adds explicit gam() and lgam() definitions, ported
from the original Cephes Math Library, so all bessel functions
in IT++ now use the proper gamma implementations.

git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1314 f17a810d-92cc-4fca-aeb8-6e32867107aa