Add cheb() functions - first order Chebyshev polynomial
The cheb() functions evaluates a first order Chebyshev polynomial
at a specific point (or set of points in a vector/matrix). Also
add a simple set of tests for the cheb() functions, and regenerate
the reference results to match the new format output in fixed
format.

Thanks to Kumar Appaiah for providing and improving the initial
patches from feature request [1869927].


git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/trunk@1398 a6533b66-300f-0410-b061-b36de6c4fb2d