Add Dolph-Chebyshev window
Add the chebwin() function to evaluate the coefficients of the
Dolph-Chebyshev window. Testing routines implemented as well.

The tests check the values output by the Dolph-Chebyshev window
function for lengths 32 and 33 for 50 dB suppression and for
lengths 127 and 128 at 25 dB suppression.

Thanks to Kumar Appaiah for providing and improving the patches.
This commit closes feature request [1869927].


git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1399 f17a810d-92cc-4fca-aeb8-6e32867107aa