Fix wrong verbose output of Parser's get() function for int and bool types
When Parser was used to scan an already defined int or bool
variable and the parsed variable was not found, the verbose
output was wrong. Here is a minimum example:

  Parser p(argc, argv);
  int i = 5;
  bool b = true;
  p.get(a, "a");
  p.get(b, "b");

The wrong output of this program was:

  i = [];
  b = ;

instead of:

  i = 5;
  b = 1;

This patch fixes this problem and also improves the parser_test
program, to detect such issues in future. Thanks to Jia-Yin for
reporting this problem.


git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/trunk@1412 a6533b66-300f-0410-b061-b36de6c4fb2d