Use zdotusub_ Fortran wrapper by default
This patch removes the run-time checks for correct zdotu_ calling
conventions, which caused a lot of portability problems in IT++
4.0.1 and 4.0.2 releases. The preferred method is now to use the
locally provided zdotusub_ Fortran wrapper (as in 4.0.0 release).
This requires the availability of a Fortran compiler, unless an
Intel MKL library is used, in which zdotu_ can be called directly
from C++ without any problems. If no Fortran compiler is available
and Intel MKL is not used, the zdotu_ function is not called at
all. In such case, a relevant warning message is printed during
the configuration step.


git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1423 f17a810d-92cc-4fca-aeb8-6e32867107aa