Removed non-portable "--with-zdotu=complex" method
It seems that there is no portable solution for calling from C++
the Fortran functions that returns a complex value. C99 introduces
"double _Complex" type, which is compatible with Fortran COMPLEX.
But this C99 type is not a valid C++ type. Therefore this patch
removes the "complex" method from "--with-zdotu" option. The valid
methods are now:
- "zdotusub" - use a zdotusub_ Fortran wrapper function
- "void" - compatible with BLAS libraries built with g77 compiler
- "none" - do not use zdotu_ function at all


git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1430 f17a810d-92cc-4fca-aeb8-6e32867107aa