Add support for "-1" indexing in Vec<>::del(i1, i2) method
The indexing operator(i1, i2) of Vec class supports "-1" indexes
denoting the last element of a vector. This feature was missing
in the del(i1, i2) method, and this patch fixes this. It also
fixes the incorrect range check (i1 <= i2).
By the way, remove the redundant local variables ii1 and ii2
from the operator(i1, i2) method.


git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1443 f17a810d-92cc-4fca-aeb8-6e32867107aa