Fix bug [1908644] in Modulator<>::set(symbols, bits2symbols) function
The bitmap matrix used in the demodulate_bits() function of the
Modulator<> base class was improperly constructed from the input
bits2symbols mapping vector. This resulted in demodulation errors
for some constellations, e.g. 64-QAM or 8-PSK. This patch fixes
this problem.

Thanks to Yann Cocheril for reporting this problem.


git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1477 f17a810d-92cc-4fca-aeb8-6e32867107aa