Fix it_assert_debug() conditions in elem_mult_out() methods of the Vec and Mat classes
The checks for equal sizes of three and more variables can not be
written like this: "a.size == b.size == c.size". The proper form
is: "(a.size == b.size) && (a.size == c.size)". This patch fixes
this bug. It also removes the redundant checks for different
dimensions of a matrix before invoking the set_size(rows, cols)
method.

Thanks to Martin Senst for reporting this bug.


git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1497 f17a810d-92cc-4fca-aeb8-6e32867107aa