Fix complex constructors in operators.cpp to build with g++ 4.3
It appears that g++ 4.3 is a bit more strict with respect to type
checking when we attempt to overload the std::complex constructor for
use with non-double types. Use of an explicit static_cast is needed to
overcome this restriction. This patch does the needful by converting
those constructors to use static_casts.

Signed-off-by: Kumar Appaiah <akumar@ee.iitm.ac.in>


git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/trunk@1510 a6533b66-300f-0410-b061-b36de6c4fb2d