Fix complex constructors in operators.cpp to build with g++ 4.3
It appears that g++ 4.3 is a bit more strict with respect to type
checking when we attempt to overload the std::complex constructor for
use with non-double types. Use of an explicit static_cast is needed to
overcome this restriction. This patch does the needful by converting
those constructors to use static_casts.

Signed-off-by: Kumar Appaiah <akumar@ee.iitm.ac.in>


git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1510 f17a810d-92cc-4fca-aeb8-6e32867107aa