Fix incorrect implementation of rint() function used under MSVC++
The rint() function is not available under MSVC++ compiler. Thus,
the local implementation has to be conditionally enabled for this
compiler. This patch fixes the incorrect implementation of rint()
function, which behaved like a similar round() function. It also
moves the "double rint(double)" definition out of the itpp
namespace.


git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1520 f17a810d-92cc-4fca-aeb8-6e32867107aa