Do not duplicate seek operation after reading data_header fields
When reading each field of a data_header structure of an IT++
file, the reading position is automatically incremented. So, the
additional seekg() operation afterwards is redundant. This small
modification fixes a problem with endless loop while reading an
itfile in programs compiled with MSVC++ 2008. Previous MSVC++
version (2005) does not show this strange behaviour.


git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1524 f17a810d-92cc-4fca-aeb8-6e32867107aa