Reset number of rows/columns if one of the dimensions given to Mat::set_size() is zero
Without this patch, the Mat::set_size() function set number of
rows or cols to non zero value, even if the other matrix
dimension was zero. This was not compatible with the constructor.
Besides, setting one of the dimensions of a matrix to non zero
value, if the other one is zero does not makes sense and might
cause problems with other functions operating on matrices. This
patch fixes this problem.

git-svn-id: https://itpp.svn.sourceforge.net/svnroot/itpp/itpp/trunk@1588 a6533b66-300f-0410-b061-b36de6c4fb2d