Fix bug [2017948] in parsing error of "a:b:c" ivec initialisation
There was a bug in vec.cpp that caused an assertion error when
integer vectors were initialized using the "a:b:c" notation. For
example: ivec iv = "0:2:4"; raised an exception. This patch fixes
this bug by removing an unnecessary "seekg" operation.

Thanks to David Hammarwall for reporting the problem and solution
to it.

git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1592 f17a810d-92cc-4fca-aeb8-6e32867107aa